services/callbacks/requirements.txt
ujson==5.2.0
paho-mqtt==1.5.1
redis==3.5.3